add some testing
